+++ app/src/main/java/com/pavelprojects/filmlibraryproject/database/FilmDatabase.kt
@@ -2,6 +2,7 @@ package com.pavelprojects.filmlibraryproject.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import com.pavelprojects.filmlibraryproject.database.dao.ChangedItemDao
 import com.pavelprojects.filmlibraryproject.database.dao.FilmItemDao
 import com.pavelprojects.filmlibraryproject.database.entity.ChangedFilmItem
 import com.pavelprojects.filmlibraryproject.database.entity.FilmItem
@@ -9,4 +10,5 @@ import com.pavelprojects.filmlibraryproject.database.entity.FilmItem
 @Database(entities = [FilmItem::class, ChangedFilmItem::class], version = 5, exportSchema = false)
 abstract class FilmDatabase : RoomDatabase() {
     abstract fun getFilmItemDao(): FilmItemDao
+    abstract fun getChangedItemDao(): ChangedItemDao
 }
