+++ app/src/main/java/com/pavelprojects/filmlibraryproject/database/dao/FilmItemDao.kt
@@ -11,37 +11,19 @@ interface FilmItemDao {
     fun insert(filmItem: FilmItem)
 
     @Insert
-    fun insert(changedFilmItem: ChangedFilmItem)
-
-    @Insert
     fun insertAll(filmList: List<FilmItem>)
-
-    @Insert
-    fun insertAllFav(changedFilmList: List<ChangedFilmItem>)
 
     @Update
     fun update(filmItem: FilmItem)
 
-    @Update
-    fun update(changedFilmItem: ChangedFilmItem)
-
     @Delete
     fun delete(filmItem: FilmItem)
-
-    @Delete(entity = ChangedFilmItem::class)
-    fun delete(changedFilmItem: ChangedFilmItem)
 
     @Query("DELETE FROM film_table")
     fun deleteAllFilms()
 
-    @Query("DELETE FROM changed_film_table")
-    fun deleteAllFavFilms()
-
     @Query("SELECT * FROM film_table")
     fun getAll(): List<FilmItem>
-
-    @Query("SELECT * FROM changed_film_table")
-    fun getAllFav(): List<FilmItem>
 
     @Query("SELECT * FROM film_table WHERE id = :id")
     fun getById(id: Long): FilmItem?
